feature/validation-handler-revamp, rearranged telegram handlers
