+++ source/group_eq_bot/telegram_handlers/messages/text/handler.py
